dbeaver/dbeaver#11684 move order and where query generation to DialectQueryGenerator class
