+++ plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/sql/SQLUtils.java
@@ -29,6 +29,7 @@ import org.jkiss.dbeaver.model.exec.DBCLogicalOperator;
 import org.jkiss.dbeaver.model.exec.DBCSession;
 import org.jkiss.dbeaver.model.impl.sql.BasicSQLDialect;
 import org.jkiss.dbeaver.model.impl.sql.RelationalSQLDialect;
+import org.jkiss.dbeaver.model.impl.sql.SQLDialectQueryGenerator;
 import org.jkiss.dbeaver.model.runtime.DBRFinder;
 import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
 import org.jkiss.dbeaver.model.runtime.VoidProgressMonitor;
@@ -58,10 +59,6 @@ public final class SQLUtils {
 
     public static final Pattern PATTERN_OUT_PARAM = Pattern.compile("((\\?)|(:[a-z0-9]+))\\s*:=");
     public static final Pattern PATTERN_SIMPLE_NAME = Pattern.compile("[a-z][a-z0-9_]*", Pattern.CASE_INSENSITIVE);
-    public static final Pattern PATTERN_COLUMN_NAME = Pattern.compile(
-        "(([a-z_][a-z0-9_]*)|(\\\"([a-z_][a-z0-9_]*)\\\"))(\\.(([a-z_][a-z0-9_]*)|(\\\"([a-z_][a-z0-9_]*)\\\")))*",
-        Pattern.CASE_INSENSITIVE
-    );
 
     private static final Pattern CREATE_PREFIX_PATTERN = Pattern.compile("(CREATE (:OR REPLACE)?).+", Pattern.CASE_INSENSITIVE | Pattern.MULTILINE);
 
@@ -468,19 +465,17 @@ public final class SQLUtils {
     {
         appendConditionString(filter, dataSource, conditionTable, query, inlineCriteria, false);
     }
-    
+
     public static void appendConditionString(
         @NotNull DBDDataFilter filter,
         @NotNull DBPDataSource dataSource,
         @Nullable String conditionTable,
-        @NotNull StringBuilder query, 
-        boolean inlineCriteria, 
+        @NotNull StringBuilder query,
+        boolean inlineCriteria,
         boolean subQuery
     ) {
-        final List<DBDAttributeConstraint> constraints = filter.getConstraints().stream()
-            .filter(x -> x.getCriteria() != null || x.getOperator() != null)
-            .collect(Collectors.toList());
-        appendConditionString(filter, constraints, dataSource, conditionTable, query, inlineCriteria, false);
+        dataSource.getSQLDialect().getQueryGenerator().appendConditionString(filter, dataSource, conditionTable, query, inlineCriteria,
+            subQuery);
     }
 
     public static void appendConditionString(
@@ -490,72 +485,9 @@ public final class SQLUtils {
         @Nullable String conditionTable,
         @NotNull StringBuilder query,
         boolean inlineCriteria,
-        boolean subQuery)
-    {
-        final String operator = filter.isAnyConstraint() ? " OR " : " AND ";  //$NON-NLS-1$ $NON-NLS-2$
-
-        for (int index = 0; index < constraints.size(); index++) {
-            final DBDAttributeConstraint constraint = constraints.get(index);
-            if (index > 0) {
-                query.append(operator);
-            }
-            if (constraints.size() > 1) {
-                // Add parenthesis for the sake of sanity
-                // Constraint may consist of several conditions and we don't want to break operator precedence
-                query.append('(');
-            }
-
-            // Attribute name could be an expression. So check if this is a real attribute
-            // and generate full/quoted name for it.
-            String attrName;
-            DBSAttributeBase cAttr = constraint.getAttribute();
-            if (cAttr instanceof DBDAttributeBinding) {
-                DBDAttributeBinding binding = (DBDAttributeBinding) cAttr;
-                if (binding.getEntityAttribute() != null &&
-                    binding.getMetaAttribute() != null &&
-                    binding.getEntityAttribute().getName().equals(binding.getMetaAttribute().getName()) ||
-                    binding instanceof DBDAttributeBindingType)
-                {
-                    if (binding.getEntityAttribute() instanceof DBSContextBoundAttribute) {
-                        DBSContextBoundAttribute entityAttribute = (DBSContextBoundAttribute) binding.getEntityAttribute();
-                        attrName = entityAttribute.formatMemberReference(true, conditionTable, DBPAttributeReferencePurpose.DATA_SELECTION);
-                    } else {
-                        attrName = DBUtils.getObjectFullName(
-                            dataSource, binding, DBPEvaluationContext.DML, DBPAttributeReferencePurpose.DATA_SELECTION
-                        );
-                    }
-                } else {
-                    if (binding.getMetaAttribute() == null || binding.getEntityAttribute() != null) {
-                        // Seems to a reference on a table column.
-                        // It is better to use real table column in expressions because aliases may not work
-                        attrName = DBUtils.getQuotedIdentifier(dataSource,
-                            subQuery ? constraint.getAttributeLabel() : constraint.getAttributeName());
-                    } else {
-                        // Most likely it is an expression so we don't want to quote it
-                        attrName = binding.getMetaAttribute().getName();
-                    }
-                }
-            } else if (cAttr != null) {
-                attrName = DBUtils.getObjectFullName(
-                    dataSource, cAttr, DBPEvaluationContext.DML, DBPAttributeReferencePurpose.DATA_SELECTION
-                );
-            } else {
-                attrName = DBUtils.getQuotedIdentifier(dataSource, constraint.getAttributeName());
-            }
-            
-            query.append(attrName).append(' ').append(getConstraintCondition(dataSource, constraint, conditionTable, inlineCriteria));
-            if (constraints.size() > 1) {
-                query.append(')');
-            }
-        }
-
-        if (!CommonUtils.isEmpty(filter.getWhere())) {
-            if (constraints.size() > 0) {
-                query.append(operator).append('(').append(filter.getWhere()).append(')');
-            } else {
-                query.append(filter.getWhere());
-            }
-        }
+        boolean subQuery) {
+        dataSource.getSQLDialect().getQueryGenerator().appendConditionString(filter, constraints, dataSource,
+            conditionTable, query, inlineCriteria, subQuery);
     }
 
     public static void appendOrderString(
@@ -563,189 +495,17 @@ public final class SQLUtils {
         @NotNull DBPDataSource dataSource,
         @Nullable String conditionTable,
         boolean subQuery,
-        @NotNull StringBuilder query)
-    {
-        // Construct ORDER BY
-        boolean hasOrder = false;
-        for (DBDAttributeConstraint co : filter.getOrderConstraints()) {
-            if (hasOrder) query.append(',');
-            String orderString = null;
-            if (co.isPlainNameReference() || co.getAttribute() == null || co.getAttribute() instanceof DBDAttributeBindingMeta || co.getAttribute() instanceof DBDAttributeBindingType) {
-                String orderColumn = subQuery ? co.getAttributeLabel() : co.getAttributeName();
-                if (canOrderByName(dataSource, co, orderColumn) && !filter.hasNameDuplicates(orderColumn)) {
-                    DBSAttributeBase attr = co.getAttribute();
-                    if (attr instanceof DBDAttributeBinding
-                        && ((DBDAttributeBinding) attr).getEntityAttribute() instanceof DBSContextBoundAttribute
-                    ) {
-                        DBDAttributeBinding attrBinding = (DBDAttributeBinding) attr;
-                        DBSContextBoundAttribute entityAttribute = (DBSContextBoundAttribute) attrBinding.getEntityAttribute();
-                        orderString = entityAttribute.formatMemberReference(
-                            true, conditionTable, DBPAttributeReferencePurpose.DATA_SELECTION
-                        );
-                    } else {
-                        // It is a simple column.
-                        orderString = co.getFullAttributeName();
-                        if (conditionTable != null) {
-                            orderString = conditionTable + '.' + orderString;
-                        }
-                    }
-                }
-            }
-            if (orderString == null) {
-                // Use position number
-                int orderIndex = getConstraintOrderIndex(filter, co);
-                if (orderIndex == -1) {
-                    log.debug("Can't generate column order: no name and no position found");
-                    continue;
-                }
-                orderString = String.valueOf(orderIndex);
-            }
-            query.append(orderString);
-            if (co.isOrderDescending()) {
-                query.append(" DESC"); //$NON-NLS-1$
-            }
-            hasOrder = true;
-        }
-        if (!CommonUtils.isEmpty(filter.getOrder())) {
-            if (hasOrder) query.append(',');
-            query.append(filter.getOrder());
-        }
-    }
-
-    private static boolean canOrderByName(@NotNull DBPDataSource dataSource, @NotNull DBDAttributeConstraint constraint, @NotNull String constraintName) {
-        if (constraint.getAttribute() == null) {
-            return true;
-        }
-        if (!dataSource.getSQLDialect().supportsOrderByIndex()) {
-            return true;
-        }
-        return PATTERN_COLUMN_NAME // we should assume columns of composite type like comp.x
-            .matcher(constraintName)
-            .matches();
+        @NotNull StringBuilder query) {
+        dataSource.getSQLDialect()
+            .getQueryGenerator()
+            .appendOrderString(filter, dataSource, conditionTable, subQuery, query);
     }
 
     @Nullable
     public static String getConstraintCondition(@NotNull DBPDataSource dataSource, @NotNull DBDAttributeConstraint constraint, @Nullable String conditionTable, boolean inlineCriteria) {
-        String criteria = constraint.getCriteria();
-        if (!CommonUtils.isEmpty(criteria)) {
-            final char firstChar = criteria.trim().charAt(0);
-            if (!Character.isLetter(firstChar) && firstChar != '=' && firstChar != '>' && firstChar != '<' && firstChar != '!') {
-                return '=' + criteria;
-            } else {
-                return criteria;
-            }
-        } else if (constraint.getOperator() != null) {
-            DBCLogicalOperator operator = constraint.getOperator();
-            StringBuilder conString = new StringBuilder();
-            Object value = constraint.getValue();
-            if (DBUtils.isNullValue(value)) {
-                if (operator.getArgumentCount() == 0) {
-                    return operator.getExpression();
-                }
-                conString.append("IS ");
-                if (constraint.isReverseOperator()) {
-                    conString.append("NOT ");
-                }
-                conString.append("NULL");
-                return conString.toString();
-            }
-            if (constraint.isReverseOperator()) {
-                conString.append("NOT ");
-            }
-            if (operator.getArgumentCount() > 0) {
-
-                if (operator.equals(DBCLogicalOperator.EQUALS) && value instanceof Object[]) {
-                    // Special case for multiple values for IN
-                    // Generate series of ORed conditions
-                    Object[] array = ((Object[]) value);
-                    for (int i = 0; i < array.length; i++) {
-                        if (i > 0) {
-                            conString.append(" OR");
-                            conString.append(' ').append(DBUtils.getQuotedIdentifier(dataSource,
-                                constraint.getAttributeLabel())).append(' ');
-                        }
-                        conString.append(operator.getExpression());
-                        String strValue = getStringValue(dataSource, constraint, inlineCriteria, array[i]);
-                        conString.append(' ').append(strValue);
-                    }
-                } else {
-                    conString.append(operator.getExpression());
-                    for (int i = 0; i < operator.getArgumentCount(); i++) {
-                        if (i > 0) {
-                            conString.append(" AND");
-                        }
-                        String strValue = getStringValue(dataSource, constraint, inlineCriteria, value);
-                        conString.append(' ').append(strValue);
-                    }
-                }
-            } else if (operator.getArgumentCount() < 0) {
-                // Multiple arguments
-                int valueCount = Array.getLength(value);
-                boolean hasNull = false, hasNotNull = false;
-                for (int i = 0; i < valueCount; i++) {
-                    final boolean isNull = DBUtils.isNullValue(Array.get(value, i));
-                    if (isNull && !hasNull) {
-                        hasNull = true;
-                    }
-                    if (!isNull && !hasNotNull) {
-                        hasNotNull = true;
-                    }
-                }
-                if (!hasNotNull) {
-                    return "IS NULL";
-                }
-                if (hasNull) {
-                    conString.append("IS NULL OR ");
-                    DBSAttributeBase attr = constraint.getAttribute();
-                    if (attr instanceof DBDAttributeBinding
-                        && ((DBDAttributeBinding) attr).getEntityAttribute() instanceof DBSContextBoundAttribute
-                    ) {
-                        DBDAttributeBinding attrBinding = (DBDAttributeBinding) attr;
-                        DBSContextBoundAttribute entityAttribute = (DBSContextBoundAttribute) attrBinding.getEntityAttribute();
-                        conString.append(entityAttribute.formatMemberReference(
-                            true, conditionTable, DBPAttributeReferencePurpose.DATA_SELECTION
-                        ));
-                    } else {
-                        if (constraint.getEntityAlias() != null) {
-                            conString.append(constraint.getEntityAlias()).append('.');
-                        } else if (conditionTable != null) {
-                            conString.append(conditionTable).append('.');
-                        }
-                        conString.append(DBUtils.getObjectFullName(
-                            dataSource, constraint.getAttribute(), DBPEvaluationContext.DML, DBPAttributeReferencePurpose.DATA_SELECTION
-                        ));
-                    }
-                    conString.append(" ");
-                }
-
-                Pair<String, String> brackets = dataSource.getSQLDialect().getInClauseParentheses();
-                conString.append(operator.getExpression());
-                conString.append(' ').append(brackets.getFirst());
-                if (!value.getClass().isArray()) {
-                    value = new Object[] {value};
-                }
-                boolean hasValue = false;
-                for (int i = 0; i < valueCount; i++) {
-                    Object itemValue = Array.get(value, i);
-                    if (DBUtils.isNullValue(itemValue)) {
-                        continue;
-                    }
-                    if (hasValue) {
-                        conString.append(",");
-                    }
-                    hasValue = true;
-                    if (inlineCriteria) {
-                        conString.append(convertValueToSQL(dataSource, constraint.getAttribute(), itemValue));
-                    } else {
-                        conString.append(dataSource.getSQLDialect().getTypeCastClause(constraint.getAttribute(), "?", true));
-                    }
-                }
-                conString.append(brackets.getSecond());
-            }
-            return conString.toString();
-        } else {
-            return null;
-        }
+        return dataSource.getSQLDialect().getQueryGenerator().getConstraintCondition(dataSource, constraint,
+            conditionTable,
+            inlineCriteria);
     }
 
     private static String getStringValue(@NotNull DBPDataSource dataSource, @NotNull DBDAttributeConstraint constraint, boolean inlineCriteria, Object value) {
@@ -1284,23 +1044,12 @@ public final class SQLUtils {
         return CommonUtils.escapeIdentifier(table.getName());
     }
 
-    public static void appendQueryConditions(DBPDataSource dataSource, @NotNull StringBuilder query, @Nullable String tableAlias, @Nullable DBDDataFilter dataFilter)
-    {
-        if (dataFilter != null && dataFilter.hasConditions()) {
-            query.append("\nWHERE "); //$NON-NLS-1$
-            appendConditionString(dataFilter, dataSource, tableAlias, query, true);
-        }
+    public static void appendQueryConditions(DBPDataSource dataSource, @NotNull StringBuilder query, @Nullable String tableAlias, @Nullable DBDDataFilter dataFilter) {
+        dataSource.getSQLDialect().getQueryGenerator().appendQueryConditions(dataSource, query, tableAlias, dataFilter);
     }
 
-    public static void appendQueryOrder(DBPDataSource dataSource, @NotNull StringBuilder query, @Nullable String tableAlias, @Nullable DBDDataFilter dataFilter)
-    {
-        if (dataFilter != null) {
-            // Construct ORDER BY
-            if (dataFilter.hasOrdering()) {
-                query.append("\nORDER BY "); //$NON-NLS-1$
-                appendOrderString(dataFilter, dataSource, tableAlias, false, query);
-            }
-        }
+    public static void appendQueryOrder(DBPDataSource dataSource, @NotNull StringBuilder query, @Nullable String tableAlias, @Nullable DBDDataFilter dataFilter) {
+        dataSource.getSQLDialect().getQueryGenerator().appendQueryOrder(dataSource, query, tableAlias, dataFilter);
     }
 
     public static boolean isExecQuery(@NotNull SQLDialect dialect, String query) {
